requirements.txt
click==6.7
dnspython==1.15.0
docker==2.1.0
docker-pycreds==0.2.1
Flask==0.12.1
Flask-HTTPAuth==3.2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
MicroFlack-Common==0.1
PyJWT==1.4.2
python-dotenv==0.6.4
python-etcd==0.4.5
requests==2.13.0
six==1.10.0
urllib3==1.20
websocket-client==0.40.0
Werkzeug==0.12.1
